Greatly simplify config validation
