+++ package.json
@@ -32,7 +32,8 @@
     "tmp": "^0.2.1",
     "ts-transformer-keys": "^0.4.3",
     "winston": "^3.7.2",
-    "yargs": "^17.5.1"
+    "yargs": "^17.5.1",
+    "zod": "^3.17.3"
   },
   "devDependencies": {
     "@types/eslint": "^8.4.3",
@@ -56,6 +57,7 @@
     "prettier": "^2.7.1",
     "ts-jest": "^28.0.5",
     "ts-node": "^10.8.1",
+    "ts-to-zod": "^1.11.0",
     "typescript": "^4.7.4"
   }
 }
