+++ src/config/config.ts
@@ -4,10 +4,60 @@
  */
 
 import { cosmiconfigSync } from "cosmiconfig";
+import { ZodError, ZodIssue } from "zod";
+import { ConfigValidationError } from "../errors";
 
 import { CopyrightConfig } from "../types";
-import ConfigBuilder from "./config-builder";
+import { clientDefinedCopyrightConfigSchema } from "./client-types";
 
+/**
+ * Format a ZodError into a human readable message.
+ *
+ * @param error The error from the zod validation
+ * @returns Some human readable string
+ */
+function formatError(error: ZodError) {
+  function formatIssue(issue: ZodIssue): string {
+    const friendlyPath = ["config", issue.path].join(".");
+
+    return `Issue: ${issue.code}\n  -> Message: ${issue.message}\n  -> Object path: ${friendlyPath}\n`;
+  }
+
+  const messageString =
+    error.issues.length === 1
+      ? "Found an error"
+      : `Found ${error.issues.length} errors`;
+
+  return `${messageString}: \n${error.issues.map(formatIssue).join("\n")}`;
+}
+
+/**
+ * The resolveConfig function just FINDS the config. The buildConfig
+ * actually does the validation
+ *
+ * @param config
+ * @returns The final config object. This is only returned if the passed config is valid.
+ * @throws {ConfigValidationError|Error}
+ *    - Config validation error happens when the config is wrong.
+ *    - Error happens when the programming is wrong.
+ */
+export function buildConfig(config: unknown): CopyrightConfig {
+  try {
+    return clientDefinedCopyrightConfigSchema.parse(config);
+  } catch (error) {
+    if (error instanceof ZodError) {
+      throw new ConfigValidationError(formatError(error));
+    }
+
+    throw Error(`Something unexpected went wrong! Error: ${error}`);
+  }
+}
+
+/**
+ * Find one of the copyright config files and build/validate it
+ *
+ * @returns The fully built copyright configuration object.
+ */
 export function resolveConfig(): CopyrightConfig {
   const explorer = cosmiconfigSync("copyright", {
     searchPlaces: ["package.json", ".copyrightrc.js", "copyright.config.js"],
@@ -21,7 +71,7 @@ export function resolveConfig(): CopyrightConfig {
     );
   }
 
-  const config = ConfigBuilder.buildConfig(result.config);
+  const config = buildConfig(result.config);
   return config;
 }
